refactor: Cleanup
